Camera2: fix warnings
CAF removed 2 aosp strings which produces of course warnings.

Remove them in all files. Seems to be that we are finished now.
TODO only to enable several qcom specific features in the PIE
UI so that the user has access.

Other then that all seems to be functional with our features and
googles UI/UX pimped by us